+++ app/src/main/java/com/vladimirpetrovski/currencyconverter/domain/usecase/RecalculateRatesUseCase.kt
@@ -4,6 +4,7 @@ import com.vladimirpetrovski.currencyconverter.domain.CalculateRatesHelper.calcu
 import com.vladimirpetrovski.currencyconverter.domain.model.CalculatedRate
 import com.vladimirpetrovski.currencyconverter.domain.repository.RatesRepository
 import io.reactivex.Single
+import java.math.BigDecimal
 import javax.inject.Inject
 
 /**
@@ -15,7 +16,7 @@ class RecalculateRatesUseCase @Inject constructor(
     private val ratesRepository: RatesRepository
 ) {
 
-    operator fun invoke(amount: Double): Single<List<CalculatedRate>> {
+    operator fun invoke(amount: BigDecimal): Single<List<CalculatedRate>> {
         val list = calculate(
             ratesRepository.cachedCalculatedRates,
             amount,
